Rename sequent
