+++ backend/src/main/kotlin/kalkulierbar/sequent/fo/FirstOrderSequentState.kt
@@ -1,4 +1,4 @@
-package kalkulierbar.sequent.psc
+package kalkulierbar.sequent.fo
 
 import kalkulierbar.sequent.GenericSequentCalculusState
 import kalkulierbar.sequent.TreeNode
@@ -6,13 +6,13 @@ import kalkulierbar.tamperprotect.ProtectedState
 import kotlinx.serialization.Serializable
 
 @Serializable
-class PSCState(
+class FirstOrderSequentState(
     override val tree: MutableList<TreeNode> = mutableListOf(),
     override var showOnlyApplicableRules: Boolean = false
 ) : GenericSequentCalculusState, ProtectedState() {
     override var seal = ""
 
     override fun getHash(): String {
-        return "psc|${tree.joinToString()}|$showOnlyApplicableRules"
+        return "fosc|${tree.joinToString()}|$showOnlyApplicableRules"
     }
 }
