+++ backend/src/main/kotlin/kalkulierbar/sequent/prop/PropositionalSequentState.kt
@@ -1,4 +1,4 @@
-package kalkulierbar.sequent.fosc
+package kalkulierbar.sequent.prop
 
 import kalkulierbar.sequent.GenericSequentCalculusState
 import kalkulierbar.sequent.TreeNode
@@ -6,13 +6,13 @@ import kalkulierbar.tamperprotect.ProtectedState
 import kotlinx.serialization.Serializable
 
 @Serializable
-class FOSCState(
+class PropositionalSequentState(
     override val tree: MutableList<TreeNode> = mutableListOf(),
     override var showOnlyApplicableRules: Boolean = false
 ) : GenericSequentCalculusState, ProtectedState() {
     override var seal = ""
 
     override fun getHash(): String {
-        return "fosc|${tree.joinToString()}|$showOnlyApplicableRules"
+        return "psc|${tree.joinToString()}|$showOnlyApplicableRules"
     }
 }
