+++ backend/src/test/kotlin/kalkulierbar/firstordersequent/TestExLeft.kt
@@ -1,15 +1,15 @@
-package kalkulierbar.fosc
+package kalkulierbar.firstordersequent
 
 import kalkulierbar.IllegalMove
 import kalkulierbar.parsers.FirstOrderParser
 import kalkulierbar.sequent.ExLeft
-import kalkulierbar.sequent.fosc.FOSC
+import kalkulierbar.sequent.fo.FirstOrderSequent
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
 class TestExLeft {
-    val instance = FOSC()
+    val instance = FirstOrderSequent()
     val parser = FirstOrderParser()
     private val instTerm = "a"
