+++ backend/src/test/kotlin/kalkulierbar/firstordersequent/TestNotLeft.kt
@@ -1,18 +1,18 @@
-package kalkulierbar.fosc
+package kalkulierbar.firstordersequent
 
 import kalkulierbar.IllegalMove
 import kalkulierbar.logic.Relation
 import kalkulierbar.parsers.FirstOrderParser
 import kalkulierbar.sequent.NotLeft
 import kalkulierbar.sequent.NotRight
-import kalkulierbar.sequent.fosc.FOSC
+import kalkulierbar.sequent.fo.FirstOrderSequent
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
 class TestNotLeft {
-    val instance = FOSC()
+    val instance = FirstOrderSequent()
     val parser = FirstOrderParser()
 
     @Test
