+++ backend/src/test/kotlin/kalkulierbar/propsequent/TestAndLeft.kt
@@ -1,18 +1,18 @@
-package kalkulierbar.psc
+package kalkulierbar.propsequent
 
 import kalkulierbar.IllegalMove
 import kalkulierbar.logic.Var
 import kalkulierbar.parsers.PropositionalParser
 import kalkulierbar.sequent.AndLeft
 import kalkulierbar.sequent.NotRight
-import kalkulierbar.sequent.psc.PSC
+import kalkulierbar.sequent.prop.PropositionalSequent
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
 class TestAndLeft {
-    val instance = PSC()
+    val instance = PropositionalSequent()
     val parser = PropositionalParser()
 
     @Test
